feat: bump go.mongodb.org/mongo-driver from 1.0.4 to 1.1.0

+++ session_test.go
@@ -114,7 +114,7 @@ func TestSession_Collection(t *testing.T) {
 
 func TestSession_SetPoolLimit(t *testing.T) {
 	type args struct {
-		limit uint16
+		limit uint64
 	}
 	tests := []struct {
 		name string
